Added Cmake definitions to allow specifying compiler
